Fixes #8690. Move langchain4j-agent-ql4j to native mode support
  - Promote langchain4j-agent-ql4j from integration-tests-jvm to
    integration-tests with @QuarkusIntegrationTest for native mode
  - Add SupportQuarkusLangchain4jProcessor build step to register
    QuarkusJsonCodecFactory classes for native reflection
  - Add GraalVM substitution for OllamaStreamingChatLanguageModel to
    work around Mutiny Context.put(String,Object) vs put(Object,Object)
    binary incompatibility between quarkus-langchain4j 1.10.0 and
    Quarkus 3.36.1 (temporary, remove once ql4j is built with Quarkus 3.36+)
  - Use @Identifier("ragAugmentor") instead of @LookupUnlessProperty to
    prevent QL4J auto-discovery of RetrievalAugmentor bean
  - Remove unused markAiServicesAsUnremovable build step
  - Add test to test-categories.yaml (group-02)
