--- integration-tests-jvm/langchain4j-agent-ql4j/README.adoc
@@ -1,34 +1,0 @@
-== Limitations of Quarkus LangChain4j being on the classpath
-
- * Does not work in the native
- * QL4J requires a different httpClient - `io.quarkiverse.langchain4j.jaxrsclient.JaxRsHttpClientBuilderFactory`. See processor and recorder for more information.
- * Even if tests are using `AgentWithoutMemory`, QL4J creates agentWithMemory automatically
-
-[IMPORTANT]
-If a `RetrievalAugmentor` with the default scope exists, QL4J detects it and applies it to the `AiAgent` (which should not have the `RetrievalAugmentor` at all).
-The issue can be simulated on the test `simpleUserMessage` (see QL4J https://github.com/quarkiverse/quarkus-langchain4j/blob/8cae9528c12bd0dce143fe2f63daa1660bfcd018/core/runtime/src/main/java/io/quarkiverse/langchain4j/RegisterAiService.java#L136[code]).
-Fot that reason, the  `RetrievalAugmentor` in `AgentProducer` is annotated via `@LookupUnlessProperty`, to disable lookup for the QL4J tests.
-Camel-langchain4j works event if lookup is disabled!
-
-== Camel Quarkus Langchain4j Agent with Quarkus LangChain4j Integration Tests
-
-By default, the langchain4j-agent-ql4j integration tests use WireMock to stub Ollama API interactions.
-
-To run the `camel-quarkus-langchain4j-agent-ql4j` integration tests against the real API, you need a Ollama instance running with the `orca-mini` & `granite4:3b` models downloaded.
-
-The MCP client tests require https://nodejs.org/[Node.js] to be installed on the test host.
-
-When Ollama is running, set the following environment variables:
-
-[source,shell]
-----
-export LANGCHAIN4J_OLLAMA_BASE_URL=your-ollama-api-url
-----
-
-If the WireMock stub recordings need updating, then remove the existing files from `src/test/resources/mappings` and run tests with either:
-
-System property `-Dwiremock.record=true`
-
-Or
-
-Set environment variable `WIREMOCK_RECORD=true`
