--- integration-tests-jvm/langchain4j-agent-ql4j/src/test/java/org/apache/camel/quarkus/component/langchain4j/agent/it/Langchain4jAgentQl4jTest.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.quarkus.component.langchain4j.agent.it;
-
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.TestProfile;
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import static org.hamcrest.Matchers.*;
-
-@ExtendWith(Langchain4jTestWatcher.class)
-@QuarkusTestResource(Langchain4jAgentTestResource.class)
-@TestProfile(Langchain4jAgentQl4jProfile.class)
-@QuarkusTest
-@Disabled("https://github.com/apache/camel-quarkus/issues/8412")
-class Langchain4jAgentQl4jTest {
-    @Test
-    void simpleUserMessage() {
-        RestAssured.given()
-                .body(Langchain4jAgentTest.TEST_USER_MESSAGE_SIMPLE)
-                .post("/langchain4j-agent/simple")
-                .then()
-                .statusCode(200)
-                .body(
-                        not(Langchain4jAgentTest.TEST_USER_MESSAGE_SIMPLE),
-                        containsString("Apache Camel"));
-    }
-}
