+++ integration-tests/langchain4j-agent-ql4j/src/main/resources/application.properties
@@ -14,6 +14,8 @@
 ## See the License for the specific language governing permissions and
 ## limitations under the License.
 ## ---------------------------------------------------------------------------
+quarkus.native.resources.includes=rag/*
+quarkus.http.test-timeout=120S
 
 quarkus.devservices.enabled=false
 quarkus.langchain4j.devservices.preload=false
