+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-268e30aa-a1a4-4ab7-bc7f-4acc98f3782c.json
