+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-28670195-38f5-4048-8148-d18b0ca75b0f.json
