+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-2919fbfe-5488-4fa3-a8f3-f64d75409b97.json
