+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-4436ba7a-8252-4ac6-8440-ead8be646a4c.json
