+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-471aa6db-7a4d-4172-8e7a-0d256117ee02.json
