+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-5262411c-d7c1-467e-b932-2aad788e6bbe.json
