+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-61945e16-28a5-4ba2-b298-ac6e7e68c5f0.json
