+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-699b4d40-169b-4ca9-92c5-5ebb653f6160.json
