+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-6ace5c02-b0f7-4368-87d4-be9c5d38dde2.json
