+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-6e571bc3-8fc5-4909-aafe-1ebc67150e0e.json
