+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-8e71c064-289e-466c-8a3d-5e3ade0be61c.json
