+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-a339a757-7f86-466a-a67d-3f519751fdd1.json
