+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-a8a0aff5-e041-4ad6-9e55-af7592b21784.json
