+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-a8d343e0-1f0d-4827-9736-76d28f52d0a5.json
