+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-df766850-0561-4506-bd7c-55eacf12dbf4.json
