+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-e45a0476-1655-4602-b20d-30332bd754bf.json
