+++ integration-tests/langchain4j-agent-ql4j/src/test/resources/mappings/agent/api_chat-f76a470e-9d73-4c08-83e8-40c371e1713e.json
