+++ tooling/scripts/test-categories.yaml
@@ -49,6 +49,7 @@ group-02:
   - kafka-oauth
   - keycloak
   - langchain4j-agent
+  - langchain4j-agent-ql4j
   - oaipmh
   - ocsf
   - pubnub
